Christian Morrow 2024-12-20
Added components:
client-form
cleint-list

Added models:
client

Added services:
client

+++ src/app/client-list/client-list.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { ClientService } from '../services/client.service';
 import { Client } from '../models/client.model';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-client-list',
+  standalone: true,
+  imports: [CommonModule],
   templateUrl: './client-list.component.html',
   styleUrls: ['./client-list.component.css']
 })
@@ -18,7 +21,12 @@ export class ClientListComponent implements OnInit {
     });
   }
 
-  deleteClient(id: number) {
+  deleteClient(id?: number) {
+    if (id === undefined) {
+      console.error('Cannot delete client: ID is undefined');
+      return;
+    }
+
     this.clientService.deleteClient(id).subscribe(() => {
       this.clients = this.clients.filter(client => client.id !== id);
     });
